+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #39 Flag to explicitly disable OpenAPI endpoints
 * #40 Flag to explicitly disable Swagger console
 * #43 Support for ZoneOffset as an API parameter
+* #44 OpenAPI model - sort schemas alphabetically
 
 ## 3.0.M1